Multivariate Normal Random Deviates using Sparse Precision

This function provides a random number generator for the multivariate normal distribution with mean equal to mean and sparse precision matrix Q.

Usage

rmvnorm_prec(Q, n = 1, mean = rep(0, nrow(Q)))

Arguments

Q

sparse precision (inverse-covariance) matrix.

n

number of observations.

mean

mean vector.

Value

a matrix with dimension length(mean) by n, containing realized draws from the specified mean and precision