summarize tinyVAST

summarize parameters from a fitted tinyVAST

Usage

# S3 method for class 'tinyVAST'
summary(object, what = c("sem", "dsem", "fixed"), ...)

Arguments

object

Output from tinyVAST()

what

What component to summarize, whether sem, dsem, or fixed for the fixed effects included in the GAM formula

...

Not used

Details

tinyVAST includes three components: a structural equation model (SEM), a dynamic structural equation model (DSEM), and a generalized additive model (GAM). Each of these are summarized and interpreted differently, and summary.tinyVAST facilitates this.

Regarding the DSEM componennt, tinyVAST includes an "arrow and lag" notation, which specifies the set of path coefficients and exogenous variance parameters to be estimated. Function tinyVAST then estimates the maximum likelihood value for those coefficients and parameters by maximizing the log-marginal likelihood.

However, many users will want to associate individual parameters and standard errors with the path coefficients that were specified using the "arrow and lag" notation. This task is complicated in models where some path coefficients or variance parameters are specified to share a single value a priori, or were assigned a name of NA and hence assumed to have a fixed value a priori (such that these coefficients or parameters have an assigned value but no standard error). The summary function therefore compiles the MLE for coefficients (including duplicating values for any path coefficients that assigned the same value) and standard error estimates, and outputs those in a table that associates them with the user-supplied path and parameter names. It also outputs the z-score and a p-value arising from a two-sided Wald test (i.e. comparing the estimate divided by standard error against a standard normal distribution).