Make a RAM (Reticular Action Model)

make_eof_ram converts SEM arrow notation to ram describing SEM parameters

Usage

make_eof_ram(
  times,
  variables,
  n_eof,
  remove_na = TRUE,
  standard_deviations = "unequal"
)

Arguments

times

A character vector listing the set of times in order

variables

A character vector listing the set of variables

n_eof

Number of EOF modes of variability to estimate

remove_na

Boolean indicating whether to remove NA values from RAM (default) or not. remove_NA=FALSE might be useful for exploration and diagnostics for advanced users

standard_deviations

One of "equal", "unequal", or a numeric vector indicating fixed values.

Value

A reticular action module (RAM) describing dependencies

Examples

# Two EOFs for two variables
make_eof_ram( times = 2010:2020, variables = c("pollock","cod"), n_eof=2 )
#> $model
#>      to  from parameter
#> 1  2010 EOF_1         1
#> 2  2011 EOF_1         2
#> 3  2012 EOF_1         3
#> 4  2013 EOF_1         4
#> 5  2014 EOF_1         5
#> 6  2015 EOF_1         6
#> 7  2016 EOF_1         7
#> 8  2017 EOF_1         8
#> 9  2018 EOF_1         9
#> 10 2019 EOF_1        10
#> 11 2020 EOF_1        11
#> 12 2010 EOF_2        NA
#> 13 2011 EOF_2        12
#> 14 2012 EOF_2        13
#> 15 2013 EOF_2        14
#> 16 2014 EOF_2        15
#> 17 2015 EOF_2        16
#> 18 2016 EOF_2        17
#> 19 2017 EOF_2        18
#> 20 2018 EOF_2        19
#> 21 2019 EOF_2        20
#> 22 2020 EOF_2        21
#> 
#> $ram
#>        heads to from parameter start
#>   [1,]     1  3    1         1  0.01
#>   [2,]     1  4    1         2  0.01
#>   [3,]     1  5    1         3  0.01
#>   [4,]     1  6    1         4  0.01
#>   [5,]     1  7    1         5  0.01
#>   [6,]     1  8    1         6  0.01
#>   [7,]     1  9    1         7  0.01
#>   [8,]     1 10    1         8  0.01
#>   [9,]     1 11    1         9  0.01
#>  [10,]     1 12    1        10  0.01
#>  [11,]     1 13    1        11  0.01
#>  [12,]     1 16    1         1  0.01
#>  [13,]     1 17    1         2  0.01
#>  [14,]     1 18    1         3  0.01
#>  [15,]     1 19    1         4  0.01
#>  [16,]     1 20    1         5  0.01
#>  [17,]     1 21    1         6  0.01
#>  [18,]     1 22    1         7  0.01
#>  [19,]     1 23    1         8  0.01
#>  [20,]     1 24    1         9  0.01
#>  [21,]     1 25    1        10  0.01
#>  [22,]     1 26    1        11  0.01
#>  [23,]     1  4    2        12  0.01
#>  [24,]     1  5    2        13  0.01
#>  [25,]     1  6    2        14  0.01
#>  [26,]     1  7    2        15  0.01
#>  [27,]     1  8    2        16  0.01
#>  [28,]     1  9    2        17  0.01
#>  [29,]     1 10    2        18  0.01
#>  [30,]     1 11    2        19  0.01
#>  [31,]     1 12    2        20  0.01
#>  [32,]     1 13    2        21  0.01
#>  [33,]     1 17    2        12  0.01
#>  [34,]     1 18    2        13  0.01
#>  [35,]     1 19    2        14  0.01
#>  [36,]     1 20    2        15  0.01
#>  [37,]     1 21    2        16  0.01
#>  [38,]     1 22    2        17  0.01
#>  [39,]     1 23    2        18  0.01
#>  [40,]     1 24    2        19  0.01
#>  [41,]     1 25    2        20  0.01
#>  [42,]     1 26    2        21  0.01
#>  [43,]     1  3   14         1  0.01
#>  [44,]     1  4   14         2  0.01
#>  [45,]     1  5   14         3  0.01
#>  [46,]     1  6   14         4  0.01
#>  [47,]     1  7   14         5  0.01
#>  [48,]     1  8   14         6  0.01
#>  [49,]     1  9   14         7  0.01
#>  [50,]     1 10   14         8  0.01
#>  [51,]     1 11   14         9  0.01
#>  [52,]     1 12   14        10  0.01
#>  [53,]     1 13   14        11  0.01
#>  [54,]     1 16   14         1  0.01
#>  [55,]     1 17   14         2  0.01
#>  [56,]     1 18   14         3  0.01
#>  [57,]     1 19   14         4  0.01
#>  [58,]     1 20   14         5  0.01
#>  [59,]     1 21   14         6  0.01
#>  [60,]     1 22   14         7  0.01
#>  [61,]     1 23   14         8  0.01
#>  [62,]     1 24   14         9  0.01
#>  [63,]     1 25   14        10  0.01
#>  [64,]     1 26   14        11  0.01
#>  [65,]     1  4   15        12  0.01
#>  [66,]     1  5   15        13  0.01
#>  [67,]     1  6   15        14  0.01
#>  [68,]     1  7   15        15  0.01
#>  [69,]     1  8   15        16  0.01
#>  [70,]     1  9   15        17  0.01
#>  [71,]     1 10   15        18  0.01
#>  [72,]     1 11   15        19  0.01
#>  [73,]     1 12   15        20  0.01
#>  [74,]     1 13   15        21  0.01
#>  [75,]     1 17   15        12  0.01
#>  [76,]     1 18   15        13  0.01
#>  [77,]     1 19   15        14  0.01
#>  [78,]     1 20   15        15  0.01
#>  [79,]     1 21   15        16  0.01
#>  [80,]     1 22   15        17  0.01
#>  [81,]     1 23   15        18  0.01
#>  [82,]     1 24   15        19  0.01
#>  [83,]     1 25   15        20  0.01
#>  [84,]     1 26   15        21  0.01
#>  [85,]     2  1    1         0  1.00
#>  [86,]     2  2    2         0  1.00
#>  [87,]     2  3    3        22    NA
#>  [88,]     2  4    4        22    NA
#>  [89,]     2  5    5        22    NA
#>  [90,]     2  6    6        22    NA
#>  [91,]     2  7    7        22    NA
#>  [92,]     2  8    8        22    NA
#>  [93,]     2  9    9        22    NA
#>  [94,]     2 10   10        22    NA
#>  [95,]     2 11   11        22    NA
#>  [96,]     2 12   12        22    NA
#>  [97,]     2 13   13        22    NA
#>  [98,]     2 14   14         0  1.00
#>  [99,]     2 15   15         0  1.00
#> [100,]     2 16   16        23    NA
#> [101,]     2 17   17        23    NA
#> [102,]     2 18   18        23    NA
#> [103,]     2 19   19        23    NA
#> [104,]     2 20   20        23    NA
#> [105,]     2 21   21        23    NA
#> [106,]     2 22   22        23    NA
#> [107,]     2 23   23        23    NA
#> [108,]     2 24   24        23    NA
#> [109,]     2 25   25        23    NA
#> [110,]     2 26   26        23    NA
#> 
#> $variances
#>        to    from parameter
#> 1   EOF_1   EOF_1         0
#> 2   EOF_2   EOF_2         0
#> 3 pollock pollock        22
#> 4     cod     cod        23
#> 
#> $standard_deviations
#> [1] "unequal"
#> 
#> attr(,"class")
#> [1] "eof_ram"