Control parameters for tinyVAST

Control parameters for tinyVAST

Usage

tinyVASTcontrol(
  nlminb_loops = 1,
  newton_loops = 0,
  eval.max = 1000,
  iter.max = 1000,
  getsd = TRUE,
  silent = getOption("tinyVAST.silent", TRUE),
  trace = getOption("tinyVAST.trace", 0),
  verbose = getOption("tinyVAST.verbose", FALSE),
  profile = c(),
  tmb_par = NULL,
  gmrf_parameterization = c("separable", "projection"),
  estimate_delta0 = FALSE,
  getJointPrecision = FALSE
)

Arguments

nlminb_loops

Integer number of times to call stats::nlminb().

newton_loops

Integer number of Newton steps to do after running stats::nlminb().

eval.max

Maximum number of evaluations of the objective function allowed. Passed to control in stats::nlminb().

iter.max

Maximum number of iterations allowed. Passed to control in stats::nlminb().

getsd

Boolean indicating whether to call TMB::sdreport()

silent

Disable terminal output for inner optimizer?

trace

Parameter values are printed every trace iteration for the outer optimizer. Passed to control in stats::nlminb().

verbose

Output additional messages about model steps during fitting?

profile

Parameters to profile out of the likelihood (this subset will be appended to random with Laplace approximation disabled).

tmb_par

list of parameters for starting values, with shape identical to tinyVAST(...)$internal$parlist

gmrf_parameterization

Parameterization to use for the Gaussian Markov random field, where the default separable constructs a full-rank and separable precision matrix, and the alternative projection constructs a full-rank and IID precision for variables over time, and then projects this using the inverse-cholesky of the precision, where this projection allows for rank-deficient covariance.

estimate_delta0

Estimate a delta model?

getJointPrecision

whether to get the joint precision matrix. Passed to sdreport.