Control parameters for tinyVAST

Control parameters for tinyVAST

Usage

tinyVASTcontrol(
  nlminb_loops = 1,
  newton_loops = 0,
  eval.max = 1000,
  iter.max = 1000,
  getsd = TRUE,
  silent = getOption("tinyVAST.silent", TRUE),
  trace = getOption("tinyVAST.trace", 0),
  verbose = getOption("tinyVAST.verbose", FALSE),
  profile = c(),
  tmb_par = NULL,
  tmb_map = NULL,
  gmrf_parameterization = c("separable", "projection"),
  reml = FALSE,
  getJointPrecision = FALSE,
  calculate_deviance_explained = TRUE,
  run_model = TRUE,
  suppress_nlminb_warnings = TRUE,
  get_rsr = FALSE,
  extra_reporting = FALSE
)

Arguments

nlminb_loops

Integer number of times to call stats::nlminb().

newton_loops

Integer number of Newton steps to do after running stats::nlminb().

eval.max

Maximum number of evaluations of the objective function allowed. Passed to control in stats::nlminb().

iter.max

Maximum number of iterations allowed. Passed to control in stats::nlminb().

getsd

Boolean indicating whether to call TMB::sdreport()

silent

Disable terminal output for inner optimizer?

trace

Parameter values are printed every trace iteration for the outer optimizer. Passed to control in stats::nlminb().

verbose

Output additional messages about model steps during fitting?

profile

Parameters to profile out of the likelihood (this subset will be appended to random with Laplace approximation disabled).

tmb_par

list of parameters for starting values, with shape identical to tinyVAST(...)$internal$parlist

tmb_map

input passed to TMB::MakeADFun as argument map, over-writing the version tinyVAST(...)$tmb_inputs$tmb_map and allowing detailed control over estimated parameters (advanced feature)

gmrf_parameterization

Parameterization to use for the Gaussian Markov random field, where the default separable constructs a full-rank and separable precision matrix, and the alternative projection constructs a full-rank and IID precision for variables over time, and then projects this using the inverse-cholesky of the precision, where this projection allows for rank-deficient covariance.

reml

Logical: use REML (restricted maximum likelihood) estimation rather than maximum likelihood? Internally, this adds the fixed effects to the list of random effects to integrate over.

getJointPrecision

whether to get the joint precision matrix. Passed to sdreport.

calculate_deviance_explained

whether to calculate proportion of deviance explained. See deviance_explained()

run_model

whether to run the model of export TMB objects prior to compilation (useful for debugging)

suppress_nlminb_warnings

whether to suppress uniformative warnings from nlminb arising when a function evaluation is NA, which are then replaced with Inf and avoided during estimation

get_rsr

Experimental option, whether to report restricted spatial regression (RSR) adjusted estimator for covariate responses

extra_reporting

Whether to report a much larger set of quantities via obj$env$report()

Value

An object (list) of class tinyVASTcontrol, containing either default or updated values supplied by the user for model settings